fix: refactor graphback to hide unused features
