+++ services/iot-graphql-api/package.json
@@ -16,9 +16,9 @@
   },
   "license": "Apache 2.0",
   "dependencies": {
-    "@graphback/runtime-knex": "~0.16.1",
-    "@graphql-codegen/add": "~2.0.1",
-    "@graphql-codegen/typescript": "~1.17.9",
+    "@graphback/runtime-knex": "0.16.2",
+    "@graphql-codegen/add": "2.0.1",
+    "@graphql-codegen/typescript": "1.17.9",
     "apollo-server-express": "~2.17.0",
     "cors": "~2.8.5",
     "dotenv": "~8.2.0",
@@ -27,20 +27,20 @@
     "graphback": "~0.16.2",
     "graphql": "~15.3.0",
     "graphql-config": "~3.0.3",
-    "graphql-migrations": "~0.16.1",
-    "graphql-subscriptions": "~1.1.0",
-    "graphql-tag": "~2.11.0",
-    "knex": "~0.21.5",
-    "pg": "~8.3.3"
+    "graphql-migrations": "0.16.2",
+    "graphql-subscriptions": "1.1.0",
+    "graphql-tag": "2.11.0",
+    "knex": "0.21.5",
+    "pg": "8.3.3"
   },
   "devDependencies": {
-    "@graphql-cli/codegen": "~1.17.8",
+    "@graphql-cli/codegen": "1.17.8",
     "@types/bson": "~4.0.2",
     "@types/glob": "~7.1.3",
-    "@types/graphql": "~14.2.3",
+    "@types/graphql": "14.5.0",
     "@types/node-fetch": "~2.5.7",
-    "graphback-cli": "~0.16.1",
-    "graphql-cli": "~4.0.0",
+    "graphback-cli": "0.16.2",
+    "graphql-cli": "4.1.0-beta.1",
     "prettier": "~2.1.1",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.3",
