+++ services/iot-graphql-api/src/index.ts
@@ -1,19 +1,14 @@
-import { ApolloServer } from 'apollo-server-express'
 import cors from 'cors'
 import express from 'express'
 import http from 'http'
-import { buildGraphbackAPI } from 'graphback'
-import { loadDBConfig, connectDB } from './db'
-import { migrateDB, removeNonSafeOperationsFilter } from 'graphql-migrations'
-import { createKnexDbProvider } from '@graphback/runtime-knex'
-import { loadConfigSync } from 'graphql-config'
 import { HTTP_RESPONSE_DELAY, NODE_ENV, HTTP_PORT } from './config'
+import { createApolloServer } from './graphback'
 
 const app = express()
 
 app.use(cors())
 
-app.use((req, res, next) => {
+app.use((req: Express.Request, res: Express.Response, next) => {
   if (HTTP_RESPONSE_DELAY && NODE_ENV === 'dev') {
     // Simulate request processing time in dev mode
     setTimeout(() => next(), HTTP_RESPONSE_DELAY)
@@ -22,43 +17,12 @@ app.use((req, res, next) => {
   }
 })
 
-const graphbackExtension = 'graphback'
-const config = loadConfigSync({
-  extensions: [
-    () => ({
-      name: graphbackExtension
-    })
-  ]
-})
-
-const projectConfig = config.getDefault()
-const graphbackConfig = projectConfig.extension(graphbackExtension)
-const modelDefs = projectConfig.loadSchemaSync(graphbackConfig.model)
-
-const db = connectDB()
-const dbConfig = loadDBConfig()
-
-const { typeDefs, resolvers, contextCreator } = buildGraphbackAPI(modelDefs, {
-  dataProviderCreator: createKnexDbProvider(db)
-})
-
-migrateDB(dbConfig, typeDefs, {
-  operationFilter: removeNonSafeOperationsFilter
-}).then(() => {
-  console.log('Migrated database')
-})
-
-const apolloServer = new ApolloServer({
-  typeDefs,
-  resolvers: [resolvers],
-  context: contextCreator
-})
-
+const apolloServer = createApolloServer();
 apolloServer.applyMiddleware({ app })
 
 const httpServer = http.createServer(app)
 apolloServer.installSubscriptionHandlers(httpServer)
 
 httpServer.listen({ port: HTTP_PORT }, () => {
-  console.log(`🚀  Server ready at http://localhost:${HTTP_PORT}/graphql`)
+  console.log(`🚀  Graphback Server ready at http://localhost:${HTTP_PORT}/graphql \nFor more information see: https://graphback.dev`)
 })
